requirements.txt
streamlit
transformers
torch
sentencepiece
fitz
PyMuPDF
python-docx
python-dotenv
google-cloud-texttospeech